pico/src/adc: Fix DMA block IRQ handler
Need to reset the write address register, as it's not the start
address register, but the actual target write address register
incrementing on each transfer.

Clear the DMA IRQ and don't reset DMA inside the IRQ handler.

Move resetting the DMA to Update() method to prevent race conditions
with IRQ handler.

Signed-off-by: Patrick Rudolph <rudolphpatrick05@gmail.com>