adc: Fix IRQ registered on wrong core
This caused the core to not wake often enough causing a buffer underflow.
That in turn causes peaks in the signal as the inverting and non-inverting
channel might get flipped.

Now the the correct core registers the IRQ handler the correct core is also
woken up on IRQ and can process the data.

Signed-off-by: Patrick Rudolph <rudolphpatrick05@gmail.com>