pico/src: Revert packet transmission after 3xh packet exchange
Tests showed that the P1P2 control unit doesn't monitor the bus and
starts transmitting after a fixed time period. This leads to bus
collisions.

Work around this send packets instead of the "correct" answer to a
3xh packet request. The remote waits up to 180msec for the "correct"
answer to be received, giving enough free bus time to transmit a packet
and receive a response.

Tests also showed that 5msec between packets is sufficient, maybe even
less silence on the bus is okay.

Signed-off-by: Patrick Rudolph <rudolphpatrick05@gmail.com>